feat: add feature infographics resident must select

+++ database/migrations/2025_09_06_154434_create_infographics_resident_must_select_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('infographics_resident_must_select', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('year')->length(4);
+            $table->integer('year')->length(4);
             $table->integer('total')->length(20);
             $table->timestamps();
         });
