+++ database/migrations/2025_09_06_160301_create_infographics_stunting_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('infographics_stunting', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('year')->length(4);
+            $table->integer('year')->length(4);
             $table->bigInteger('stunting_id')->length(20)->index();
             $table->bigInteger('total')->length(20);
             $table->boolean('is_active')->default(true);
